+++ amr-wind/physics/mms/MMS.cpp
@@ -1,4 +1,4 @@
-#include "amr-wind/mms/MMS.H"
+#include "amr-wind/physics/mms/MMS.H"
 #include "amr-wind/CFDSim.H"
 #include "AMReX_ParmParse.H"
 #include "AMReX_iMultiFab.H"
