+++ unit_tests/mms/test_mms_init.cpp
@@ -2,7 +2,7 @@
 #include "aw_test_utils/iter_tools.H"
 #include "aw_test_utils/test_utils.H"
 
-#include "amr-wind/mms/MMS.H"
+#include "amr-wind/physics/mms/MMS.H"
 
 namespace amr_wind_tests {
 TEST_F(MMSMeshTest, mms_initialization)
