+++ unit_tests/mms/test_mms_src.cpp
@@ -2,8 +2,8 @@
 #include "aw_test_utils/iter_tools.H"
 #include "aw_test_utils/test_utils.H"
 
-#include "amr-wind/mms/MMS.H"
-#include "amr-wind/mms/MMSForcing.H"
+#include "amr-wind/physics/mms/MMS.H"
+#include "amr-wind/physics/mms/MMSForcing.H"
 #include "amr-wind/equation_systems/icns/icns.H"
 #include "amr-wind/equation_systems/icns/icns_ops.H"
